Refactor tests by module and share fixtures
